nit: create module cli_wrappers
